Refactored code to remove duplication and improve readabiliy
