+++ react-dapp/src/Election.tsx
@@ -18,41 +18,3 @@ export const privKey = new paillierBigint.PrivateKey(
   pubKey
 );
 export const maxVotes = 9;
-export function tallyVotes(votes: string[]) {
-  let voteResult = BigInt(votes[0]);
-
-  for (let i = 1; i < votes.length; i++) {
-    voteResult = pubKey.addition(voteResult, BigInt(votes[i]));
-  }
-  console.log(voteResult);
-  return voteResult;
-}
-
-export function decryptTotal(
-  privKey: paillierBigint.PrivateKey,
-  encryptedTotal: bigint,
-  numOptions: number
-) {
-  let decryptedResult = privKey.decrypt(encryptedTotal).toString();
-  while (decryptedResult.length < maxVotes * numOptions)
-    decryptedResult = "0" + decryptedResult;
-  console.log(decryptedResult);
-  return decryptedResult;
-}
-
-export function decodeResult(encodedResult: string) {
-  const decodedResult = encodedResult
-    .toString()
-    .match(/\d{1,9}/g)
-    ?.map((x) => +x.toString());
-  decodedResult?.reverse();
-  return decodedResult;
-}
-
-export function handleRevert(err: unknown) {
-  if (err instanceof Error) {
-    console.log(err.message);
-    const message = err.message.match(/"message":"(.*?)"/)![0].slice(11, -1);
-    alert(message.replace("execution reverted: ", ""));
-  }
-}
