+++ react-dapp/src/pages/Voting.css
